+++ assets/props/switches/lever.lua
@@ -1,0 +1,12 @@
+isEnabled = false
+
+function onInteract(selfEty, userEty)
+  isEnabled = not isEnabled
+  if isEnabled then
+    setAnimation(selfEty, "on")
+    modifyWirePower(selfEty, 1)
+  else
+    setAnimation(selfEty, "off")
+    modifyWirePower(selfEty, -1)
+  end
+end