+++ library/Kathu/World/WorldSpace.hs
@@ -7,12 +7,13 @@ module Kathu.World.WorldSpace where
 import           Apecs                     hiding (Map)
 import qualified Apecs
 import           Apecs.Physics             hiding (Map)
-import           Control.Lens              hiding ((.=), set)
+import           Control.Lens              hiding ((.=))
 import           Control.Monad             (foldM)
 import           Control.Monad.IO.Class    (MonadIO)
 import           Data.Aeson
 import           Data.Aeson.Types          (Parser, typeMismatch)
 import qualified Data.Map                  as Map
+import           Data.Maybe                (fromMaybe)
 import           Data.Text                 (Text)
 import qualified Data.Text                 as T
 import           Data.Vector               (Vector)
@@ -69,6 +70,7 @@ data WorldSpace = WorldSpace
     , _worldEntities      :: !(Vector InstancedPrefab)
     , _worldItems         :: !(Vector InstancedItem)
     , _worldChunks        :: !Chunks
+    , _worldInventory     :: !Identifier
     }
 makeLenses ''WorldSpace
 
@@ -77,7 +79,14 @@ instance Monoid WorldSpace  where mempty = emptyWorldSpace
 instance Component WorldSpace  where type Storage WorldSpace  = Global WorldSpace
 
 emptyWorldSpace :: WorldSpace
-emptyWorldSpace = WorldSpace "" "the void..." "" Map.empty (V2 0 0) False Map.empty Nothing Vec.empty Vec.empty emptyChunks
+emptyWorldSpace = WorldSpace "" "the void..." "" Map.empty (V2 0 0) False Map.empty Nothing Vec.empty Vec.empty emptyChunks "void"
+
+-- | A collections of obtained items that are specific to certain worlds
+newtype WorldInventory = WorldInventory {unWorldInventory :: IDMap (IDMap ItemStack)}
+
+instance Semigroup WorldInventory  where (<>) = mappend
+instance Monoid WorldInventory  where mempty = WorldInventory Map.empty
+instance Component WorldInventory  where type Storage WorldInventory  = Global WorldInventory
 
 -------------------
 -- Serialization --
@@ -123,7 +132,7 @@ instance (FromJSON (Dependency s m ItemStack), Functor m) => FromJSON (Dependenc
 ----------------
 
 encodeValueForWorldSpace :: MonadIO m => AllTiles' -> WorldSpace -> m Value
-encodeValueForWorldSpace allTiles (WorldSpace wid wname initPal palettes loadPos shouldSavePos vars script etys items chunks) = do
+encodeValueForWorldSpace allTiles (WorldSpace wid wname initPal palettes loadPos shouldSavePos vars script etys items chunks inventory) = do
     fieldPairs <- serializeFieldSetPairs allTiles chunks
     pure . object $
         [ "world-id"                   .= wid
@@ -133,6 +142,7 @@ encodeValueForWorldSpace allTiles (WorldSpace wid wname initPal palettes loadPos
         , "palettes"                   .= Map.keys palettes
         , "load-point"                 .= (loadPos / unitsPerTile)
         , "should-save-exact-position" .= shouldSavePos
+        , "inventory-id"               .= inventory
         , "variables"                  .= vars
         , "entities"                   .= Vec.reverse etys  -- as part of loading this get reversed, so we un-reverse it once we save to ensure the same order
         , "items"                      .= Vec.reverse items
@@ -147,12 +157,13 @@ instance ( s `CanProvide` IDMap Prefab
          , MonadIO m
          ) => FromJSON (Dependency s m WorldSpace) where
     parseJSON (Object v) = do
-        worldId    <- v .: "world-id"
+        wID        <- v .: "world-id"
         wName      <- v .: "name"
         initPal    <- v .: "initial-palette"
         paletteIDs <- v .: "palettes"
         loadPnt    <- (*unitsPerTile) <$> v .: "load-point"
         variables  <- v .:? "variables" .!= Map.empty
+        worldInvId <- fromMaybe wID <$> v .:? "inventory-id"
 
         shouldSerializePosition <- v .:? "should-save-exact-position" .!= False
         chunkConfigs :: Vector ChunkConfig <- v .: "chunks"
@@ -178,8 +189,8 @@ instance ( s `CanProvide` IDMap Prefab
                 liftDependency $
                     foldM (applyChunkConfig legend) emptyChunks chunkConfigs
 
-        pure $ WorldSpace worldId wName initPal <$> palettes <*> pure loadPnt <*> pure shouldSerializePosition <*> pure variables <*> wScript <*> entities <*> items <*> newChunks
-    parseJSON e          = typeMismatch "WorldSpace" e
+        pure $ WorldSpace wID wName initPal <$> palettes <*> pure loadPnt <*> pure shouldSerializePosition <*> pure variables <*> wScript <*> entities <*> items <*> newChunks <*> pure worldInvId
+    parseJSON e = typeMismatch "WorldSpace" e
 
 -- | When we serialize the WorldSpace, we serialize the fields in the following order
 worldspaceFieldOrder :: FieldOrder
@@ -191,6 +202,7 @@ worldspaceFieldOrder = mkFieldOrderFromList
     , "palettes"
     , "load-point"
     , "should-save-exact-position"
+    , "inventory-id"
     , "variables"
     , "legend"
     , "chunks"