+++ verda/ChangeLog.md
@@ -9,3 +9,4 @@ Initial merger of verda game engine. New features are as follows:
 * Texture scaling can be set on a per file basis with @nearest or @linear file name suffixes
 Issues to resolve before merging to master:
 * Palettes no longer work, need to implement shader behavior for this
+* Allow pixels per unit to be set